requirements.txt
asgiref==3.7.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
djangorestframework==3.14.0
idna==3.4
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pytesseract==0.3.10
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==2.0.2
